backend/requirements.txt
flask>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
psycopg2-binary>=2.9.0
requests>=2.28.0
tzdata>=2024.1; platform_system == "Windows"
